+++ projects/igniteui-angular/src/lib/grids/filtering/grid-filtering-row.component.ts
@@ -101,10 +101,6 @@ export class IgxGridFilteringRowComponent implements AfterViewInit {
         this.filter();
     }
 
-    get locale() {
-        return window.navigator.language;
-    }
-
     @ViewChild('defaultFilterUI', { read: TemplateRef })
     protected defaultFilterUI: TemplateRef<any>;