+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes for each version of this project will be documented in this 
 
 ## 7.1.1
 ### Features
-- `IgxTimePickerComponent`: in addition to the current dialog interaction mode, the user now select or edit a time value, using an editable masked input with a dropdown.
+- `IgxTimePickerComponent`: in addition to the current dialog interaction mode, now the user can select or edit a time value, using an editable masked input with a dropdown.
 
 ## 7.1.0
 ### Features